+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [v0.1.6]
 - fix application icon for windows
+- fix push notifications title
 
 ## [v0.1.1], [v0.1.2], [v0.1.3], [v0.1.4], [v0.1.5]
 - Update experiments 🧪